+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ---
 
 ## [0.1.0] - 2025-06-04 - Vlad
+### Added
+- Bot's /start command
+
 ### Updated
 - Keyword extraction